Add support for handling smaller deletions
The use case here is where the patient lacks multiple exons, such that
the proposed exon skips are actually smaller deletions than the variant
already present. In this case, we do not need to update the variants at
all, since the proposed therapy will have no effect (the exon is already
missing).