Add inverted to Variant
This is required for variants on transcripts located on the reverse
strand. In the mutalyzer delins model of these variants, the 'inverted'
key is located inside the inserted/deleted objects. If this is not set
correctly (i.e. set to True for transcripts on the reverse strand), the
protein prediction that is generated from the delins model is not
correct.